backend/requirements.txt
fastapi
uvicorn
mss
pygetwindow
pydirectinput
pillow
opencv-python
numpy
pytesseract
python-dotenv
websockets
requests
